@@ -58,8 +58,8 @@ dependencies:
   shared_preferences: ^2.0.17
   sms_otp_auto_verify: ^2.1.0
   swipe_refresh: ^1.0.2
-  url_launcher: ^6.1.8
-  webview_flutter: ^3.0.4
+  url_launcher: ^6.1.11
+  webview_flutter: ^4.4.2
   get_storage: ^2.1.1
   shimmer: ^3.0.0
   dio:
