@@ -1,22 +1,29 @@
 import 'dart:convert';
+import 'dart:developer';
 import 'package:flutter/foundation.dart';
 import 'package:get/get.dart';
+import 'package:get_storage/get_storage.dart';
 import 'package:http/http.dart' as http;
 import 'package:rentitezy/model/property_model.dart';
 import 'package:rentitezy/utils/const/app_urls.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'utils/const/appConfig.dart';
+import '../../utils/const/appConfig.dart';
+import '../../utils/services/rie_user_api_service.dart';
+import '../model/property_list_nodel.dart';
 
-class PropertyApiController extends GetxController {
+class HomeController extends GetxController {
   var isLoading = true.obs;
   var isLoadingLocation = true.obs;
-  var apiPropertyList = <PropertyModel>[].obs;
-  var allPropertyData = <PropertyModel>[].obs;
+  PropertyListModel? allPropertyData;
+  var apiPropertyList = <PropertyListModel>[].obs;
+ // var allPropertyData = <PropertyModel>[].obs;
   var selectedIndex = 0.obs;
   var locationBy = "ALL".obs;
   var categories = <String>[].obs;
   final Future<SharedPreferences> prefs = SharedPreferences.getInstance();
   var offset = 10.obs;
+
+  final RIEUserApiService apiService = RIEUserApiService();
 
   @override
   void onInit() {
@@ -42,29 +49,24 @@ class PropertyApiController extends GetxController {
     if (!isNext) {
       isLoading(true);
     }
-
-    var sharedPreferences = await prefs;
     String url = '${AppUrls.listing}?limit=10&available=true&offset=$offset';
     if (locationBy.value != 'ALL') {
       url =
           '${AppUrls.listing}?limit=10&available=true&offset=$offset&location=${locationBy.value}';
     }
 
-    final response = await http.get(
-      Uri.parse(url),
-      headers: <String, String>{
-        "Auth-Token": sharedPreferences.getString(Constants.token).toString()
-      },
-    );
-    if (response.statusCode == 200) {
-      var body = jsonDecode(response.body);
-      bool success = body["success"];
+    final response = await apiService.getApiCallWithURL(endPoint: url);
+
+     // var body = jsonDecode(response.body);
+      bool success = response["success"];
       try {
         if (success) {
-          apiPropertyList.value = (body["data"] as List)
+
+          allPropertyData = PropertyListModel.fromJson(response);
+             /* (response["data"] as List)
               .map((stock) => PropertyModel.fromJson(stock))
-              .toList();
-          allPropertyData.addAll(apiPropertyList);
+              .toList();*/
+          //allPropertyData.addAll(apiPropertyList);
         }
         if (!isNext) {
           isLoading(false);
@@ -72,18 +74,14 @@ class PropertyApiController extends GetxController {
       } catch (e) {
         debugPrint(e.toString());
       }
-    } else {
-      Get.snackbar("Error", 'Error during fetch api data');
-    }
   }
 
   void fetchAddress() async {
     isLoadingLocation(true);
-    var sharedPreferences = await prefs;
     final response = await http.get(
       Uri.parse(AppUrls.locations),
       headers: <String, String>{
-        "Auth-Token": sharedPreferences.getString(Constants.token).toString()
+        "Auth-Token": GetStorage().read(Constants.token).toString()
       },
     );
     if (response.statusCode == 200) {
@@ -107,7 +105,7 @@ class PropertyApiController extends GetxController {
 
   locationFunc(String newVal) {
     locationBy.value = newVal;
-    allPropertyData.value = [];
+    //allPropertyData.value = [];
     update();
     refresh();
     fetchProperties(false);
