@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:rentitezy/utils/const/appConfig.dart';
+import 'package:get_storage/get_storage.dart';
 import 'package:rentitezy/utils/const/app_urls.dart';
 import 'screen/splash_screen.dart';
 
 void main() {
   runApp(const MyApp());
+  GetStorage.init();
 }
 
 class MyApp extends StatelessWidget {
@@ -15,7 +16,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return GetMaterialApp(
       title: AppUrls.appName,
-      debugShowCheckedModeBanner: false,
+     // debugShowCheckedModeBanner: false,
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
