@@ -1,12 +1,13 @@
 import 'dart:convert';
 
 import 'package:get/get.dart';
+import 'package:get_storage/get_storage.dart';
 import 'package:intl/intl.dart';
 import 'package:rentitezy/utils/const/appConfig.dart';
-import 'package:rentitezy/controller.dart';
+import 'package:rentitezy/home/home_controller/home_controller.dart';
 import 'package:rentitezy/pdf/pdf_api.dart';
 import 'package:rentitezy/pdf/pdf_new.dart';
-import 'package:rentitezy/screen/home_screen.dart';
+import 'package:rentitezy/home/home_view/home_screen.dart';
 import 'package:rentitezy/widgets/const_widget.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
@@ -22,7 +23,7 @@ class MainViewController extends GetxController {
   var profileImg = 'guest'.obs;
   var isTenant = false.obs;
   final Future<SharedPreferences> pref = SharedPreferences.getInstance();
-  final propertyApiController = Get.put(PropertyApiController());
+  final propertyApiController = Get.put(HomeController());
   static MainViewController get to => Get.find();
   @override
   void onInit() {
@@ -31,11 +32,10 @@ class MainViewController extends GetxController {
   }
 
   void addToFav(String listingId) async {
-    var sharedPreferences = await pref;
     final response = await http.post(Uri.parse(AppUrls.addFav),
         headers: <String, String>{
           'Content-Type': 'application/json',
-          "Auth-Token": sharedPreferences.getString(Constants.token).toString()
+          "Auth-Token": GetStorage().read(Constants.token).toString()
         },
         body: jsonEncode(<String, String>{
           "listingId": listingId,
